requirements.txt
asgiref==3.2.10
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-cors-headers==3.5.0
django-dotenv==1.4.2
djangorestframework==3.11.1
drf-yasg==1.20.0
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
packaging==20.4
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.2
uuid==1.30
zipp==3.4.0
